+++ src/main/java/me/shedaniel/rei/client/ConfigObject.java
@@ -46,8 +46,8 @@ public class ConfigObject {
     
     public boolean preferVisibleRecipes = false;
     
-    @Comment("Enable support for old REI plugins which uses registerSpeedCraft")
-    public boolean enableLegacySpeedCraftSupport = false;
+    //    @Comment("Enable support for old REI plugins which uses registerSpeedCraft")
+    //    public boolean enableLegacySpeedCraftSupport = false;
     
     @Comment("Force enable 2019 REI April Fools' joke")
     public boolean aprilFoolsFish2019 = false;